requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
django-mongodb-backend
pymongo
python-dotenv
cloudinary
django-cloudinary-storage
djangorestframework-simplejwt
gunicorn
whitenoise
